requirements.txt
gpt4all==2.0.2
python-telegram-bot==13.7
